perf: 修改usersession 模块位置
