+++ apps/rbac/tree.py
@@ -98,7 +98,6 @@ special_pid_mapper = {
     'terminal.endpoint': 'terminal_node',
     'terminal.endpointrule': 'terminal_node',
     'audits.ftplog': 'terminal',
-    'users.usersession': 'terminal',
     'perms.view_myassets': 'my_assets',
     'ops.celerytask': 'task_center',
     'ops.view_celerytaskexecution': 'task_center',
@@ -113,6 +112,7 @@ special_pid_mapper = {
     "settings.view_setting": "view_setting",
     "rbac.view_console": "view_console",
     "rbac.view_audit": "view_audit",
+    'audits.usersession': 'view_audit',
     "rbac.view_workbench": "view_workbench",
     "rbac.view_webterminal": "view_workbench",
     "rbac.view_filemanager": "view_workbench",
