+++ apps/users/api/__init__.py
@@ -6,5 +6,4 @@ from .preference import *
 from .profile import *
 from .relation import *
 from .service import *
-from .session import *
 from .user import *
