+++ apps/users/models/__init__.py
@@ -4,6 +4,5 @@
 
 from .group import *
 from .preference import *
-from .session import *
 from .user import *
 from .utils import *
