+++ apps/users/serializers/__init__.py
@@ -4,5 +4,4 @@ from .group import *
 from .preference import *
 from .profile import *
 from .realtion import *
-from .session import *
 from .user import *
